weird ...

+++ srcs/libft.c
@@ -5,7 +5,7 @@ int   ft_abs(int x)
 	return x < 0 ? -x : x;
 }
 
-float   ft_fabs(float x)
+double   ft_fabs(double x)
 {
 	return x < 0 ? -x : x;
 }
